requirements.txt
# Core dependencies - minimal for Pi 5B
pyserial>=3.5
